[ssh] Rename ssh_utils file to ssh_signal
